==> src/3_dwc_vernacularname_mapping.Rmd <==
---
title: "Darwin Core mapping script for Vernacular Names Extension"
subtitle: "For: Inventory of alien species in Europe (DAISIE)"
author:
- Lien Reyserhove
- David Roy
date: "`r Sys.Date()`"
output:
  html_document:
    df_print: paged
    number_sections: yes
    toc: yes
    toc_depth: 3
    toc_float: yes
---

This file describes the steps required to map the data to [Vernacular Names](http://rs.gbif.org/extension/gbif/1.0/vernacularname.xml).

# Setup 

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```

Load libraries:

```{r}
library(tidyverse)      # To do data science
library(magrittr)       # To use %<>% pipes
library(janitor)        # To clean input data
library(readxl)         # To read Excel files
library(rgbif)          # To use GBIF services
```

# Read data

1. Define data types

```{r}
col_types <- cols(.default = col_character())
```

2. Read vernacular names data

```{r}
vernacular_names <- read_delim(
  file = here::here("data", "raw", "input_vernacular_names.csv"),
  delim = ",",
  col_types = col_types
)
```

3. Read literature references

```{r}
literature_references <- read_delim(
  file = here::here("data", "interim", "interim_literature_references.csv"),
  delim = ",",
  col_types = col_types
)
```

4. Read `core_taxa`

```{r}
core_taxa <-
  read_csv(here::here("data", "interim", "core_taxa.csv"), col_types = col_types)
```

# Map vernacular names extension

Preview of the data:

```{r}
vernacular_names %>% head(n = 5)
```

## taxonID

1. Remove all taxonID's that are not included in the taxon core

```{r}
vernacular_names %<>% filter(idspecies %in% core_taxa$taxonID)
```

2. Map taxonID

```{r}
vernacular_names %<>% mutate(dwc_taxonID = as.numeric(idspecies))
```

## vernacularName

```{r}
vernacular_names %<>% mutate(dwc_vernacularName = name)
```

## source

Join `vernacular_names` with `source` information in `literature_references`:

```{r}
vernacular_names %<>%
  left_join(
    select(literature_references, sourceid, source),
    by = "sourceid"
  )
```

Rename `source` to `dwc_source`

```{r}
vernacular_names %<>% rename("dwc_source" = "source")
```

## language

1. Show unique values for `language`

```{r}
vernacular_names %>%
  group_by(language) %>%
  summarize(records = n())
```

2. Map language information to [ISO 639-1 Language Codes](https://tools.gbif.org/dwca-validator/vocabulary.do?id=http://iso.org/639-1)

```{r}
vernacular_names %<>% mutate(dwc_language = recode(language,
  "Belorussian" = "be",
  "Czech" = "cs",
  "Danish" = "da",
  "Dutch" = "nl",
  "English" = "en",
  "Estonian" = "et",
  "Faeroese" = "fo",
  "Finnish" = "fi",
  "French" = "fr",
  "German" = "de",
  "Greek" = "el",
  "Hebrew (western characters)" = "he",
  "Hebrew (Hebrew characters)" = "he",
  "Hungarian" = "hu",
  "Icelandic" = "is",
  "Italian" = "it",
  "Latvian" = "lv",
  "Lithuanian" = "lt",
  "Maltese" = "mt",
  "Norwegian" = "no",
  "Polish" = "pl",
  "Portuguese" = "pt",
  "Romanian" = "ro",
  "Spanish" = "es",
  "Swedish" = "sv",
  "Turkish" = "tr"
))
```

```{r}
vernacular_names %>%
  group_by(language, dwc_language) %>%
  summarize(records = n())
```

# Post-processing

1. Only keep the Darwin Core columns:

```{r}
vernacular_names %<>% select(starts_with("dwc_"))
```

2. Drop the `dwc_` prefix:

```{r}
colnames(vernacular_names) <- str_replace(colnames(vernacular_names), "dwc_", "")
```

3. Preview data:

```{r}
vernacular_names %>% head()
```

4. Sort on `taxonID`:

```{r}
vernacular_names %<>% arrange(taxonID)
```

5. Save to CSV:

```{r}
write_csv(vernacular_names,
  here::here("data", "processed", "vernacularname.csv"),
  na = ""
)
```
